fix(@aws-amplify/analytics): resolve/reject correctly when calling reocrd
